# Conflicts:
#	cmake-build-debug/.cmake/api/v1/reply/codemodel-v2-adc65c91bf4d6eb21b37.json
#	cmake-build-debug/.cmake/api/v1/reply/target-SDL_game-Debug-d1e78b494c334d900c80.json
#	cmake-build-debug/.ninja_deps
#	cmake-build-debug/.ninja_log
#	cmake-build-debug/CMakeFiles/SDL_game.dir/engine/GameEngine.cpp.obj
#	cmake-build-debug/CMakeFiles/SDL_game.dir/game/Game.cpp.obj
#	cmake-build-debug/CMakeFiles/SDL_game.dir/game/gameObject/GameObject.cpp.obj
#	cmake-build-debug/CMakeFiles/SDL_game.dir/game/gameObject/platform/Platform.cpp.obj
#	cmake-build-debug/CMakeFiles/SDL_game.dir/game/gameObject/wall/Wall.cpp.obj
#	cmake-build-debug/CMakeFiles/SDL_game.dir/game/player/Player.cpp.obj
#	cmake-build-debug/CMakeFiles/SDL_game.dir/main.cpp.obj
#	cmake-build-debug/CMakeFiles/SDL_game.dir/vc140.pdb
#	cmake-build-debug/CMakeFiles/clion-Debug-log.txt
#	cmake-build-debug/SDL_game.exe
#	cmake-build-debug/SDL_game.ilk
#	cmake-build-debug/SDL_game.pdb
#	cmake-build-debug/Testing/Temporary/LastTest.log
#	engine/GameEngine.cpp
#	engine/GameEngine.h
#	game/Game.cpp
#	game/Game.h
#	game/gameObject/GameObject.h
#	main.cpp

+++ game/gameObject/GameObject.cpp
@@ -10,7 +10,7 @@ void GameObject::render(SDL_Renderer * render) {
     SDL_Rect rect = {int(x), int(y), int(width), int(height)};
     SDL_RenderFillRect(render, &rect);
 }
-void GameObject::move(Direction direction,float dx,float dy){
+void GameObject::move(Direction direction,int dx,int dy){
     switch (direction) {
         case Direction::LEFT:
         case Direction::RIGHT:
